first steps with mongoDB
