+++ controllers/contacts/index.js
@@ -1,0 +1,7 @@
+const getAll = require("./getAll");
+const add = require("./add");
+
+module.exports = {
+	getAll,
+	add,
+};
