+++ middlewares/index.js
@@ -1,0 +1,5 @@
+const ctrlWrapper = require("./ctrlWrapper");
+
+module.exports = {
+	ctrlWrapper,
+};
