+++ package.json
@@ -11,8 +11,10 @@
 	"dependencies": {
 		"cors": "2.8.5",
 		"cross-env": "7.0.3",
+		"dotenv": "^16.0.1",
 		"express": "4.17.1",
 		"joi": "^17.6.0",
+		"mongoose": "^6.4.0",
 		"morgan": "1.10.0",
 		"uuid": "^8.3.2"
 	},
@@ -23,6 +25,6 @@
 		"eslint-plugin-import": "2.25.3",
 		"eslint-plugin-node": "^11.1.0",
 		"eslint-plugin-promise": "5.2.0",
-		"nodemon": "2.0.15"
+		"nodemon": "^1.3.3"
 	}
 }
